Move contracts to src/moonbeam/contracts folder
