+++ package.json
@@ -7,7 +7,6 @@
     "shibuya": "dotenv -e .env babel-node src/astar/shibuya.js",
     "rocstar": "dotenv -e .env babel-node src/astar/rocstar.js",
     "shiden": "dotenv -e .env babel-node src/astar/shiden.js",
-    "setup-moonbase-alpha": "dotenv -e .env babel-node src/moonbeam/setup-moonbase-alpha.js",
     "moonbase-alpha": "dotenv -e .env babel-node src/moonbeam/moonbase-alpha.js",
     "moonbase": "dotenv -e .env babel-node src/moonbeam/moonbase.js",
     "utils": "dotenv -e .env babel-node src/utils.js",
