+++ src/moonbeam/contracts/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "box",
+  "name": "contracts",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
